clean up refactor
