+++ sdk/AGENTS.md
@@ -8,7 +8,7 @@
 - Never export the chat widget under `src/chat/` or reach it by any route other than the dynamic import in `src/client.ts`, so it stays inside the client pass's lazy chunk with its own bundled React.
 - Everything the chat widget imports must be a devDependency so tsdown inlines it into the chat chunk; adding it to `dependencies` would force hosts to install it.
 - Generate shadcn components with `deno task ui add <component>`; never edit `src/components/ui` beyond minimal typed fixups, and keep `components.json` aligned with webapp's style, base color, and icon library.
-- Hand-written chat components live one per file directly in `src/components` (shadcn keeps `src/components/ui`); shared helpers go in `src/lib/utils.ts`, constants in `src/lib/constants.ts`, shared types in `src/lib/types.ts`, and the debug logger in `src/lib/debug.ts`.
+- Hand-written chat components live one per file directly in `src/components` (shadcn keeps `src/components/ui`). `src/lib` splits along the lazy-chunk boundary: `client-types.ts`, `client-constants.ts`, and `client-utils.ts` hold what the eager client entry loads (public API types, defaults, the debug-line logger) and must not import chat-chunk modules, while `types.ts`, `constants.ts`, `utils.ts` (shared helpers, including shadcn's `cn`), and `debug.ts` (stream-chunk logging) belong to the lazy chat chunk.
 - `src/chat/styles.generated.ts` is gitignored, never edited by hand, and regenerated by `deno task generate:styles` (run first by both `build` and `check`); a fresh checkout needs one of those before a standalone `typecheck` or `knip` can resolve it.
 - Keep `examples/*` standalone consumer apps on the built `dist` via their `file:../..` dependency, with no Tailwind or shadcn of their own, so they keep demonstrating the shadow-root style boundary.
 - Publish with `deno task build` followed by `npm publish` from `sdk`; the package ships only `dist`, `README.md`, `LICENSE`, and `package.json`.
