+++ sdk/src/lib/constants.ts
@@ -1,5 +1,5 @@
-/** Endpoint the chat widget streams from when the mount options give none. */
-export const DEFAULT_ENDPOINT = "/api/chat"
+// Chat-chunk-only constants; anything the eager client entry needs lives in
+// client-constants.ts so this module never enters dist/client.js.
 
 // Client tools every chat mount declares to the agent: the endpoint forwards
 // them verbatim and the chat widget executes them in the host page.
