+++ webapp/src/routes/api/chat/-lib/types.ts
@@ -1,0 +1,6 @@
+import type { chatParamsFromRequest } from "@tanstack/ai"
+
+export type ChatParams = Awaited<ReturnType<typeof chatParamsFromRequest>>
+export type ChatMessages = ChatParams["messages"]
+
+export type DebugLog = (category: string, summary: string, data?: unknown) => void
